Adds `adaptto == "bo"` option for mesh adaptation
This option is conceptually similar to a componentwise version of the
"wavemaker" (i.e. the Frobenius norm of the structural sensitivty tensor).
In this case, it adapts the mesh to the base state "b" and the overlap "o"
between each component of the direct and adjoint eigenmodes. This should
give an economic alternative to adapting to the base, direct, and adjoint
state vectors independently. The idea is similar to the adaptation based
on structural sensitivity proposed by Fabre et al. AMR. (2018), but allows
for better adaptation on mixed spaces that are more complex than those
found in vanilla Navier-Stokes.